+++ CHANGES.md
@@ -1,3 +1,3 @@
 # 0.1.0dev (unreleased)
 
-- Create package `picarrito`
+- Create package `opentoolflux`